requirements.txt
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.35.0
llama-cpp-python>=0.2.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
